rename to MetadataSectionHelper
